skip playwright install after first time

+++ scripts/start.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-npx playwright install-deps chromium
-su node -c "npx playwright install chromium"
+
+if [ ! -f done ]; then
+  npx playwright install-deps chromium
+  su node -c "npx playwright install chromium"
+fi
+
+touch done
 su node -c "node index.js"
